+++ lib/src/taro_loader.dart
@@ -3,7 +3,6 @@ import 'dart:typed_data';
 import 'package:taro/src/taro_exception.dart';
 import 'package:taro/src/taro_loader_network.dart';
 import 'package:taro/src/taro_loader_storage.dart';
-import 'package:taro/src/taro_type.dart';
 
 /// [TaroLoader] is a class that manages different types of loaders.
 class TaroLoader {
@@ -12,7 +11,7 @@ class TaroLoader {
 
   /// The [TaroLoaderNetwork] instance used to load data from the network.
   /// Loader is able to change the original network loader.
-  TaroLoaderNetwork _networkLoader = const TaroLoaderNetwork();
+  TaroLoaderNetwork _networkLoader = TaroLoaderNetwork();
 
   /// The [TaroLoaderStorage] instance used to load data from the storage.
   /// Loader is able to change the original storage loader.
@@ -33,11 +32,19 @@ class TaroLoader {
   Future<Uint8List> load({
     required String url,
     required Map<String, String> headers,
-    required TaroHeaderOption headerOption,
+    bool checkMaxAgeIfExist = false,
+    bool ifThrowMaxAgeHeaderError = false,
+    Duration? customCacheDuration,
   }) async {
-    final storageBytes = await _storageLoader.load(
-      url: url,
-    );
+    Uint8List? storageBytes;
+    try {
+      storageBytes = await _storageLoader.load(
+        url: url,
+      );
+    } on Exception {
+      // ignore exception
+    }
+
     if (storageBytes != null) {
       return storageBytes;
     }
@@ -45,7 +52,9 @@ class TaroLoader {
     final networkResponse = await _networkLoader.load(
       url: url,
       headers: headers,
-      headerOption: headerOption,
+      checkMaxAgeIfExist: checkMaxAgeIfExist,
+      ifThrowMaxAgeHeaderError: ifThrowMaxAgeHeaderError,
+      customCacheDuration: customCacheDuration,
     );
 
     if (networkResponse == null) {
@@ -55,12 +64,16 @@ class TaroLoader {
     }
 
     // save to storage
-    await _storageLoader.save(
-      url: url,
-      bytes: networkResponse.bytes,
-      contentType: networkResponse.contentType,
-      expireAt: networkResponse.expireAt,
-    );
+    try {
+      await _storageLoader.save(
+        url: url,
+        bytes: networkResponse.bytes,
+        contentType: networkResponse.contentType,
+        expireAt: networkResponse.expireAt,
+      );
+    } on Exception {
+      // ignore exception
+    }
 
     return networkResponse.bytes;
   }